# src/main/resources/openapi/tax-entity-api.yaml
swagger: "2.0"

info:
  version: "0.0.1"
  title: Goods and Services Tax - API for a Tax Entity resource
  description: |
    Operations related to the `tax-entity` resource
  license:
    name: ""
    url: ""
  contact:
    name: Ksheerasagar Akella
    url: ""
    email: ksheerasagar.akella@gmail.com

consumes:
  - application/json
produces:
  - application/json

# Dummy host
host: tax-entity.resources.gst.com

schemes:
- http

basePath: /gst/resources

#
# Resources
#
paths:

  /v1/tax-entity:

    post:
      tags:
      - "TaxEntity"
      operationId: createResource
      description: |
        Create the resource
      parameters:
      - in: body
        name: tax-entity
        description: |
          The details of the resource to be created
        required: true
        schema:
          $ref: "#/definitions/v1.TaxEntity"
      responses:
        200:
          description: |
            *OK* - The resource was created successfully
          schema:
            $ref: "#/definitions/v1.TaxEntity"
        400:
          $ref: "#/responses/400-BadRequest"
        401:
          $ref: "#/responses/401-Unauthorized"
        403:
          $ref: "#/responses/403-Forbidden"
        404:
          $ref: "#/responses/404-NotFound"
        default:
          $ref: "#/responses/500-InternalServerError"

#
# Parameters
#
parameters:

#
# Standard REST query parameters
#


#
# Definitions
#
definitions:

  v1.TaxEntity:
    description: |
      A single `TaxEntity` resource
    type: object
    required:
    - pan
    - email
    - cellPhone
    properties:
      pan:
        description: |
          The Permanent Account Number (PAN) of the tax entity
        type: string
      email:
        description: |
          The email of the tax-entity
        type: string
      cellPhone:
        description: |
          The cell phone number of the tax-entity
        type: string

#
# Common models
#
  v1.PatchElement:
    description: |
      The 'PatchElement' object
    properties:
      op:
        description: |
          The operation of this 'PatchElement' ('replace' or 'add')
        type: string
      path:
        type: string
        description: |
          The path of this 'PatchElement' ('/status' or '/notes' or '/name' or '/activationStatus' or
          '/description')
      value:
        description: |
          The value of this 'PatchElement'
        type: object

  v1.ErrorResponse:
    type: object
    description: |
      The standard error response
    properties:
      link:
        type: string
        format: url
        description: |
          The href to the item
        readOnly: true
      rel:
        type: string
        description: |
          The relation type for the item
        readOnly: true
        default: help
      code:
        description: |
          The code associated with the error
        readOnly: true
        type: integer
      message:
        description: |
          The message associated with the error
        readOnly: true
        type: string

#
# Responses
#
responses:
  400-BadRequest:
    description: |
      *Bad Request* - client has sent malformed request:
      missing mandatory query parameters, invalid value
      in query parameters, backend exception etc.
    schema:
      $ref: "#/definitions/v1.ErrorResponse"

  401-Unauthorized:
    description: |
      *Unauthorized* - the server rejected the client's credentials
    schema:
      $ref: "#/definitions/v1.ErrorResponse"

  403-Forbidden:
    description: |
      *Forbidden* - server rejected client request operation
      that is outside of the permitted scope.
    schema:
      $ref: "#/definitions/v1.ErrorResponse"

  404-NotFound:
    description: |
      *Not Found* - server has not found the requested URI:
      missing path parameters, invalid web resource etc.
    schema:
      $ref: "#/definitions/v1.ErrorResponse"

  500-InternalServerError:
    description: |
      *Internal Server Error* - server failed in handling the
      client request.
    schema:
      $ref: "#/definitions/v1.ErrorResponse"